Reprex fixed! The issue here is that col_factor() is a function in two different packages: readr and scales. The readr function is the one you want so you can specify that with readr::col_factor() instead of col_factor().
(by default, when a function appears in two different packages that are loaded into the environment, R will use the package that was loaded in most recently).

(scales is a package that likely got loaded with another library: either ggplot2 itself or, more likely, with the plotting library used to make your nice-looking RF plot).